Move common PV serverless calls to common util
