+++ plugin-flex-ts-template-v2/src/feature-library/conference/custom-components/ConnectingParticipants/ConnectingParticipants.tsx
@@ -6,7 +6,7 @@ import { useDispatch, useSelector } from 'react-redux';
 import AppState from '../../../../types/manager/AppState';
 import { reduxNamespace } from '../../../../utils/state';
 import { removeConnectingParticipant, ConferenceState } from '../../flex-hooks/states/ConferenceSlice';
-import ConferenceService from '../../utils/ConferenceService';
+import ProgrammableVoiceService from '../../../../utils/serverless/ProgrammableVoice/ProgrammableVoiceService';
 import { FetchedCall } from '../../../../types/serverless/twilio-api';
 
 export interface OwnProps {
@@ -36,7 +36,7 @@ const ConnectingParticipants = (props: OwnProps) => {
       .filter((p) => p.conferenceSid === props.task?.conference?.conferenceSid)
       .forEach((participant) => {
         // if this call is no longer active, remove it
-        ConferenceService.getCallProperties(participant.callSid)
+        ProgrammableVoiceService.getCallProperties(participant.callSid)
           .then((response: FetchedCall) => {
             if (
               response &&
