+++ plugin-flex-ts-template-v2/src/feature-library/conference/flex-hooks/actions/UnholdParticipant.ts
@@ -1,6 +1,6 @@
 import * as Flex from '@twilio/flex-ui';
 
-import ConferenceService from '../../utils/ConferenceService';
+import ProgrammableVoiceService from '../../../../utils/serverless/ProgrammableVoice/ProgrammableVoiceService';
 import { isConferenceEnabledWithoutNativeXWT } from '../../config';
 import { FlexActionEvent, FlexAction } from '../../../../types/feature-loader';
 
@@ -20,6 +20,6 @@ export const actionHook = function handleUnholdConferenceParticipant(flex: typeo
 
     const conferenceSid = task.conference?.conferenceSid || task.attributes?.conference?.sid;
     abortFunction();
-    await ConferenceService.unholdParticipant(conferenceSid, participantSid);
+    await ProgrammableVoiceService.unholdParticipant(conferenceSid, participantSid);
   });
 };
