+++ plugin-flex-ts-template-v2/src/feature-library/conference/flex-hooks/custom-actions/startExternalWarmTransfer.ts
@@ -1,6 +1,6 @@
 import { Actions, ITask, Manager, Notifications, TaskHelper } from '@twilio/flex-ui';
 
-import ConferenceService from '../../utils/ConferenceService';
+import ProgrammableVoiceService from '../../../../utils/serverless/ProgrammableVoice/ProgrammableVoiceService';
 import { addConnectingParticipant } from '../states/ConferenceSlice';
 import { ConferenceNotification } from '../notifications/Conference';
 
@@ -29,7 +29,7 @@ export const registerStartExternalWarmTransfer = async () => {
       console.log(`Adding ${phoneNumber} to conference`);
       let participantCallSid;
       try {
-        participantCallSid = await ConferenceService.addParticipant(mainConferenceSid, callerId, phoneNumber);
+        participantCallSid = await ProgrammableVoiceService.addParticipant(mainConferenceSid, callerId, phoneNumber);
 
         Manager.getInstance().store.dispatch(
           addConnectingParticipant({
