+++ plugin-flex-ts-template-v2/src/utils/serverless/ProgrammableVoice/ProgrammableVoiceService.ts
@@ -5,7 +5,7 @@
 
 */
 
-import ApiService from '../../../utils/serverless/ApiService';
+import ApiService from '../ApiService';
 import { EncodedParams } from '../../../types/serverless';
 import { FetchedCall, FetchedConferenceParticipant } from '../../../types/serverless/twilio-api';
 
@@ -19,11 +19,70 @@ export interface ParticipantResponse {
   participantsResponse: FetchedConferenceParticipant;
 }
 
-export interface RemoveParticipantResponse {
+export interface GenericSuccessResponse {
   success: boolean;
 }
 
-class ConferenceService extends ApiService {
+class ProgrammableVoiceService extends ApiService {
+  startColdTransfer = async (callSid: string, to: string, from?: string): Promise<boolean> => {
+    const { success } = await this._startColdTransfer(callSid, to, from);
+    return success;
+  };
+
+  _startColdTransfer = async (callSid: string, to: string, from?: string): Promise<any> => {
+    const encodedParams: EncodedParams = {
+      Token: encodeURIComponent(this.manager.store.getState().flex.session.ssoTokenPayload.token),
+      callSid: encodeURIComponent(callSid),
+      to: encodeURIComponent(to),
+      from: from ? encodeURIComponent(from) : '',
+    };
+
+    return this.fetchJsonWithReject<GenericSuccessResponse>(
+      `${this.serverlessProtocol}://${this.serverlessDomain}/common/flex/programmable-voice/cold-transfer`,
+      {
+        method: 'post',
+        headers: { 'Content-Type': 'application/x-www-form-urlencoded' },
+        body: this.buildBody(encodedParams),
+      },
+    );
+  };
+
+  _toggleMuteParticipant = async (conferenceSid: string, participantSid: string, muted: boolean): Promise<string> => {
+    return new Promise((resolve, reject) => {
+      const encodedParams: EncodedParams = {
+        conference: encodeURIComponent(conferenceSid),
+        participant: encodeURIComponent(participantSid),
+        muted: encodeURIComponent(muted),
+        Token: encodeURIComponent(this.manager.user.token),
+      };
+
+      this.fetchJsonWithReject<ParticipantResponse>(
+        `${this.serverlessProtocol}://${this.serverlessDomain}/common/flex/programmable-voice/mute-participant`,
+        {
+          method: 'POST',
+          headers: { 'Content-Type': 'application/x-www-form-urlencoded' },
+          body: this.buildBody(encodedParams),
+        },
+      )
+        .then((response: ParticipantResponse) => {
+          console.log(`${muted ? 'Muted' : 'Unmuted'} successful for participant`, participantSid);
+          resolve(response.participantsResponse.callSid);
+        })
+        .catch((error) => {
+          console.error(`Error ${muted ? 'muting' : 'un-muting'} participant ${participantSid}\r\n`, error);
+          reject(error);
+        });
+    });
+  };
+
+  muteParticipant = async (conference: string, participantSid: string): Promise<string> => {
+    return this._toggleMuteParticipant(conference, participantSid, true);
+  };
+
+  unmuteParticipant = async (conference: string, participantSid: string): Promise<string> => {
+    return this._toggleMuteParticipant(conference, participantSid, false);
+  };
+
   _toggleParticipantHold = async (conference: string, participantSid: string, hold: boolean): Promise<string> => {
     return new Promise((resolve, reject) => {
       const encodedParams: EncodedParams = {
@@ -128,7 +187,7 @@ class ConferenceService extends ApiService {
         Token: encodeURIComponent(this.manager.user.token),
       };
 
-      this.fetchJsonWithReject<RemoveParticipantResponse>(
+      this.fetchJsonWithReject<GenericSuccessResponse>(
         `${this.serverlessProtocol}://${this.serverlessDomain}/common/flex/programmable-voice/remove-participant`,
         {
           method: 'POST',
@@ -136,7 +195,7 @@ class ConferenceService extends ApiService {
           body: this.buildBody(encodedParams),
         },
       )
-        .then((_response: RemoveParticipantResponse) => {
+        .then((_response: GenericSuccessResponse) => {
           console.log(`Participant ${participantSid} removed from conference`);
           resolve(participantSid);
         })
@@ -174,4 +233,4 @@ class ConferenceService extends ApiService {
   };
 }
 
-export default new ConferenceService();
+export default new ProgrammableVoiceService();
